requirements.txt
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
copier==7.2.0
cryptography==41.0.2
distro==1.8.0
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.2
funcy==2.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jsonschema==3.2.0
packaging==23.1
paramiko==3.2.0
pipx==1.1.0
pluggy==1.2.0
plumbum==1.8.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyNaCl==1.5.0
pyrsistent==0.19.3
pytest==7.4.0
python-dotenv==0.21.1
PyYAML==5.4.1
pyyaml-include==1.3
requests==2.31.0
six==1.16.0
smmap==5.0.0
texttable==1.6.7
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
userpath==1.8.0
websocket-client==0.59.0
